Add vehicle and visitor pages
